compilation

+++ utils.h
@@ -60,6 +60,6 @@ void print_statistics(long long int program_time);
  *
  * @param time_ms long sleep time in milliseconds.
  */
-void msleep(long time_ms);
+void msleep(int time_ms);
 
 #endif
